+++ apps/home/app/(profile-readme)/@readme/_shared/content/manudeli.mdx
