+++ apps/home/app/(profile-readme)/@readme/_shared/content/minsoo-web.mdx
