+++ apps/home/app/(profile-readme)/@readme/default.tsx
@@ -1,9 +1,9 @@
+import { Readme } from '@/_shared'
 import fs from 'node:fs'
 import path from 'node:path'
-import { Readme } from '@/_shared'
 
 export default function Page() {
-  const filePath = path.join(process.cwd(), 'app/@readme/_shared/content/main.mdx')
+  const filePath = path.join(process.cwd(), 'app/(profile-readme)/@readme/_shared/content/main.mdx')
   const mainReadme = fs.readFileSync(filePath, 'utf8')
 
   return <Readme text={mainReadme} />
