+++ apps/home/app/(profile-readme)/@readme/people/page.tsx
@@ -1,6 +1,6 @@
+import { Readme } from '@/_shared'
 import fs from 'node:fs'
 import path from 'node:path'
-import { Readme } from '@/_shared'
 
 export default async function Page({
   searchParams,
@@ -8,7 +8,11 @@ export default async function Page({
   searchParams: { username: string }
 }) {
   const { username } = searchParams
-  const filePath = path.join(process.cwd(), 'app/@readme/_shared/content', `${username}.mdx`)
+  const filePath = path.join(
+    process.cwd(),
+    'app/(profile-readme)/@readme/_shared/content',
+    `${username}.mdx`,
+  )
   const readme = fs.readFileSync(filePath, 'utf8')
 
   return <Readme text={readme} />
