+++ apps/home/app/(profile-readme)/people/people.constants.ts
