+++ apps/home/next.config.js
@@ -6,7 +6,6 @@ const nextConfig = {
   transpilePackages: ['@hamsurang/ui', '@hamsurang/icon'],
   experimental: {
     typedRoutes: true,
-    missingSuspenseWithCSRBailout: false,
   },
   webpack(config) {
     const fileLoaderRule = config.module.rules.find((rule) => rule.test?.test?.('.svg'))
